+++ include/service/helper.h
@@ -16,9 +16,10 @@
  * General Public License version 2 for more details.
  */
 #pragma once
+#include <nul/types.h>
 
-void * operator new[](unsigned size, unsigned alignment);
-void * operator new(unsigned size, unsigned alignment);
+void * operator new[](size_t size, unsigned alignment);
+void * operator new(size_t size, unsigned alignment);
 void  do_exit(const char *msg) __attribute__((noreturn));
 extern void *(*memalloc)(unsigned long size, unsigned long align);
 extern void (*memfree)(void *ptr);
